TS conversion: spacefill representation
