+++ src/buffer/sphere-buffer.ts
@@ -16,7 +16,7 @@ export interface SphereBufferData extends BufferData {
 const SphereBufferDefaultParameters = Object.assign({
   disableImpostor: false
 }, SphereGeometryBufferDefaultParameters)
-type SphereBufferParameters = typeof SphereBufferDefaultParameters
+export type SphereBufferParameters = typeof SphereBufferDefaultParameters
 
 /**
  * Sphere buffer. Depending on the value {@link ExtensionFragDepth} and
